[feat] Add shape types to sine.py.
The main changes are:

+ Make the class generic in a variable `DimModel` and set the type of
`self.dim_model` as `DimModel`. That way, when it is initialized with
`self.dim_model`, Pyre will store the type. We have to add
`Generic[DimModel]` in this class because `self.dim_model` is defined in
this class.

+ Annotate `forward` as accepting a Tensor of shape N x N and returning
a Tensor of shape N x N x DimModel.

This is about as straightforward as it gets for shape types. There were no
major complications in this class. If even this is too cumbersome or hard to
follow, it will not get easier in other classes :) So, feedback welcome on the
understandability.

Note:

+ I added `from __future__ import annotations` so that the type
annotations are not evaluated at runtime. This is necessary because the
runtime Tensor class is not generic and cannot take arguments like
`Tensor[torch.float32, N]`, etc.

+ I changed `torch.arange(0.0, seq_len)` to `torch.arange(0, seq_len,
dtype=torch.float32)` because Pyre won't be able to calculate the
resulting Tensor shape using floats like 0.0. It is able to do so with
integers like 0.

+ An inconvenience is that we check if x.ndim == 2 and then unsqueeze
the last dimension. However, this means that there could be two possible
shapes for x, whereas we have declared that it has exactly shape N x N.
So, I added a local variable `output` with the desired shape after
unsqueezing and suppressed the Pyre error with a comment.

I'll try to see if Pyre can allow such idioms in the future.