[fMHA] Update interface for `attn_bias` argument
We now expect bias to be in BMHK format when inputs are in BMHK format.
We also do *not* broadcast the bias dimensions, and expect the user to do it before
calling `memory_efficient_attention`. This makes it much easier to handle the backprop as well,
as we would need to handle a reduction in case of a broadcast in the FW pass.

ghstack-source-id: b3e91d61ee16be0d04337fed11a886bb983fa591
Pull Request resolved: https://github.com/fairinternal/xformers/pull/743

__original_commit__ = fairinternal/xformers@5cada9ceedd3519ddd5d91e5bb90ca453fceb88d