[fused seqpar] Support multi-query settings in Triton
In allgather+matmul we required the strides of the weight and of the output matrices to be equal. When using multi-query, wq is larger than wk/wv, thus its strides are different. (This is actually not true, because wq/wk/wv are column-major, hence their stride is in_feats, not out_feats, but the outputs are row-major thus their strides pose a problem).

The solution is just to support arbitrary strides. It's very ugly and bulky since we need to pass tons of parameters around, but it works.

ghstack-source-id: d40a452c7121fdd9992773bfc2c14aeba503804e
Pull Request resolved: https://github.com/fairinternal/xformers/pull/1007

__original_commit__ = fairinternal/xformers@cb710110ca7a309bbd39a52b9f78b3458ee3faa4