[ci] Remove timm and the test that uses it
The only reason we install torchvision in our CI is because timm needs it. But torchvision is a pain because its nightlies are not in sync with PyTorch's ones and, since old nightlies get pruned, it's often impossible to install PyTorch + torchvision nightlies.

Thus we just remove timm and all its usages in tests.

ghstack-source-id: 95072d2534a4a12b23b7e7219d6915e2323cd3df
Pull Request resolved: https://github.com/fairinternal/xformers/pull/1110

__original_commit__ = fairinternal/xformers@696ebf7a6733ac2ab6647a76de2d762761422d34