[fused seqpar] Fix race condition when mixing dtypes
In https://github.com/fairinternal/xformers/pull/1022 we made the rendezvous faster by attempting to use the Store underlying the ProcessGroup to exchange the addresses. However, the keys we used for that Store were always the same. When mixing dtypes, we create multiple instances of the _FusedSequenceParallel, and the second one's rendezvous could risk loading an old stale value from the store before it was overwritten.

ghstack-source-id: c3279d1644d53fa8d74a2e4acdbb18c59ae178e3
Pull Request resolved: https://github.com/fairinternal/xformers/pull/1106

__original_commit__ = fairinternal/xformers@83cc53dbdc4b391e37b66dce464e828bfc07d17a