Fix SAC for older PyTorch (fairinternal/xformers#1146)
* Fix SAC for older PyTorch

https://github.com/fairinternal/xformers/pull/1138 had fixed it for nightly PyTorch, at the expense of older PyTorch. Now that PyTorch added some BC functionality, we can revert some of the changes and adapt only what was needed

* More fixes

Note: I had to disable same block being recomputed within AMP, because the _copy_to operators from AMP were lost within checkpointing. Need to see if this is actually a problem or not. We could have it work before because we could just recompute the op before, but not sure if this is the behavior we will want in the future

__original_commit__ = fairinternal/xformers@e32aea111ad6c741b4f90b32b8e4b9d974e27176