Make profiler trace analysis faster
Once the profiler has run, we sum all of the flops for all the operators.
To do so, we need to de-duplicate events (eg SDPA calling MM: we should count the flops only once). When there is ambiguity, we chose to count the flops of the parent operator (if we know its flops).
This process was quite slow when we had a lot of events (~45k for a Llama7B on difformers), and could take up to 5mn (`O(N^2)` algo). Now it takes ~1sec (`O(N*ln(N))`)

ghstack-source-id: 7aac17cf85ae68990e37467c40aa0d5a89e8927c
Pull Request resolved: https://github.com/fairinternal/xformers/pull/1203

__original_commit__ = fairinternal/xformers@9820ed05a6edc3ebc0d5f314774d0f3bfaebe652