[fused seqpar] Avoid straggling kernels
This is needed for capturing a CUDA graph (i.e., all work must be "joined back" into capturing stream).

But it's also needed by the next commit in the stack to avoid a race condition between the seq-num increase kernel of the next iteration and the write_values kernel of the previous iteration.

ghstack-source-id: 4f8e9bfada42566b83b1325cb6dca0bd05374249
Pull Request resolved: https://github.com/fairinternal/xformers/pull/1034

__original_commit__ = fairinternal/xformers@c57618e3ac6e214532a145ef7fe6cf125f6636f3