[fused seqpar] Make counter-based synchronization protocol clearer
We had two bugs (deadlocks) due to incorrect updates to the shared counters that are used by processes to coordinate. This was because the same counter was updated at different times by different processes. This felt very error-prone (no clear owner for a counter) thus brittle.

I'm proposing here what I believe to be a clearer and more robust solution, which uses two clearly-defined counters:
- one which tracks the completion of full operations (waited on at the beginning, updated at the end)
- one which updates mid-operation by the producer rank to signal to the consumer rank that it's ready (either it sent the data, if push-based, or it readied the data, if pull-based)

This protocol works! The tests all pass. Except... the ones based on Triton, because Triton expects the counters to be a contiguous tensor, and this PR breaks that. This can be fixed but it's too much effort for now. Hence I'm just parking this PR here as a reference for the future.

ghstack-source-id: 112aa146768b4e0bd1c22afef23778f15eb29278
Pull Request resolved: https://github.com/fairinternal/xformers/pull/989

__original_commit__ = fairinternal/xformers@1b629075cd47a437d527ff8c7492267b2638f069