Use a constant with clearly-defined type for log2e in fwd_kernel_splitK (#1181)
Summary:
Triton 3.2 made some changes to its interpretation of constants
(https://github.com/triton-lang/triton/pull/4613) which makes Triton more
consistent with pytorch/numpy, but cause some surprising issues with this
kernel.  Specifically it seems like log2e is interpreted as float32 in one
instance and float64 in another, which leads to reduced prediction accuracy in
some cases.

To prevent this, let's make log2e a constant and define it as float32.